Refactor archive location and login style
