--- server/src/main/resources/static/styles/pages/login-style.scss
@@ -1,148 +1,0 @@
-* {
-	margin: 0;
-	padding: 0;
-	box-sizing: border-box;
-	font-family: 'Arial', sans-serif;
-}
-
-img {
-	position: relative;
-	left: 0;
-	width: 400px;
-	z-index: 10;
-	padding: -50px 0;
-	opacity: 0;
-	transform: translateY(20px);
-	animation: fadeUp 1s forwards;
-}
-
-@keyframes fadeUp {
-	to {
-		opacity: 1;
-		transform: translateY(0);
-	}
-}
-
-body {
-	position: relative;
-	background-color: #ffffff;
-	display: flex;
-	flex-direction: column;
-	justify-content: flex-start;
-	align-items: center;
-	min-height: 100vh;
-	padding: 50px 20px 20px;
-}
-
-.container {
-	width: 100%;
-	max-width: 800px;
-	min-width: 600px;
-	max-height: 600px;
-	min-height: 400px;
-	border-radius: 12px;
-	overflow: hidden;
-	box-shadow: 0 6px 18px rgba(0, 0, 0, 0.15);
-	border: 2px solid #8CD790;
-	opacity: 0;
-	transform: translateY(20px);
-	animation: fadeUp 1s forwards;
-}
-
-.tab-container {
-	display: flex;
-	width: 100%;
-}
-
-.tab {
-	flex: 1;
-	padding: 15px;
-	text-align: center;
-	font-weight: bold;
-	border: none;
-}
-
-.tab-login {
-	background-color: #8CD790;
-	color: white;
-}
-
-.tab-register {
-	background-color: #e0e0e0;
-	color: #333;
-}
-
-.formulario {
-	padding: 32px;
-	background-color: white;
-}
-
-.form-group {
-	margin-bottom: 24px;
-}
-
-label {
-	display: block;
-	margin-bottom: 6px;
-	font-weight: 500;
-	color: #333;
-}
-
-.password-container {
-	position: relative;
-}
-
-input {
-	width: 100%;
-	padding: 14px 16px;
-	border: 1px solid #8CD790;
-	border-radius: 6px;
-	font-size: 18px;
-	outline: none;
-}
-
-input:focus {
-	border-color: #8CD790;
-	box-shadow: 0 0 0 2px rgba(140, 215, 144, 0.25);
-}
-
-input[type="password"] {
-	padding-right: 40px;
-}
-
-.button-container {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	margin-top: 16px;
-}
-
-.primary-button {
-	background-color: #8CD790;
-	color: white;
-	border: none;
-	padding: 14px 28px;
-	margin-bottom: -80px;
-	border-radius: 4px;
-	font-weight: 500;
-	font-size: 18px;
-	cursor: pointer;
-}
-
-.primary-button:hover {
-	background-color: #7BC680;
-}
-
-.forgot-password {
-	color: blue;
-	text-decoration: underline;
-	font-size: 14px;
-	background: none;
-	border: none;
-	cursor: pointer;
-	margin-bottom: -80px;
-}
-
-.forgot-password:hover {
-	color: #050036;
-}
