+++ server/src/main/resources/templates/pages/access/login.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="es"
+	  layout:decorate="~{/layouts/basic-layout}"
+	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+	  xmlns:th="http://www.thymeleaf.org">
+<head>
+	<meta charset="UTF-8">
+	<meta content="width=device-width, initial-scale=1.0" name="viewport">
+	<link rel="stylesheet" th:href="@{/styles/pages/access.css}">
+	<title>Login</title>
+</head>
+
+<body layout:fragment="body-content">
+
+<a class="kn-logo">
+	<img alt="Logo de Knowy" class="kn-logo--img" src="/images/logoWoutText.png">
+	<h1 class="kn-logo--text">KNOWY</h1>
+</a>
+
+<div class="kn-container">
+	<div class="tab-container">
+		<a class="tab-container__tab tab-container--active" href=#>Iniciar Sesión</a>
+		<a class="tab-container__tab tab-container--inactive" href=#>Registrarse</a>
+	</div>
+
+	<form action="#" class="kn-form" method="post" th:action="@{/login}" th:object="${loginForm}">
+		<div class="kn-form__group">
+			<label class="kn-form--label" for="email">Correo o usuario</label>
+			<input class="kn-form--input" id="email" name="email" placeholder="Correo o usuario" th:field="*{email}"/>
+		</div>
+
+		<div class="form-group">
+			<label class="kn-form--label" for="password">Contraseña</label>
+			<input class="kn-form--input" id="password" name="password" th:field="*{password}"
+				   type="password"/>
+		</div>
+
+		<button class="kn-button" type="submit">Iniciar Sesión</button>
+		<a class="forgot-password" type="button">¿Has olvidado la contraseña?</a>
+	</form>
+</div>
+
+</body>
+</html>
