+++ server/src/main/resources/templates/pages/access/register.html
@@ -44,8 +44,7 @@
 			<input class="kn-form--input" id="repeat-password" type="password">
 		</div>
 
-		<button class="kn-button" type="submit">Iniciar Sesión</button>
-		<a class="forgot-password">¿Has olvidado la contraseña?</a>
+		<button class="kn-button" type="submit">Registrarse</button>
 	</form>
 </div>
 </body>
