--- server/src/main/resources/templates/pages/login-index.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="es"
-	  xmlns:th="http://www.thymeleaf.org"
-	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<link rel="stylesheet" th:href="@{/styles/pages/login-style.css}">
-	<title>Login</title>
-</head>
-<body>
-<img src="/images/knowylogo.png" alt="Logo de Knowy">
-<div class="container">
-	<div class="tab-container">
-		<div class="tab tab-login"><a href=# style="text-decoration: none; color: white">Iniciar Sesión</a></div>
-		<div class="tab tab-register"><a href=# style="text-decoration: none; color:black">Registrarse</a></div>
-	</div>
-
-	<div class="formulario">
-		<form action="#" th:action="@{/login}" method="post" th:object="${loginForm}">
-			<div class="form-group">
-				<label for="email">Correo o usuario</label>
-				<input id="email" name="email" placeholder="Correo o usuario" th:field="*{email}" />
-			</div>
-			<div class="form-group">
-				<label for="password">Contraseña</label>
-				<div class="password-container">
-					<input type="password" id="password" name="password" placeholder="***************" th:field="*{password}" />
-				</div>
-			</div>
-			<div class="button-container">
-				<button type="submit" class="primary-button">Iniciar</button>
-				<button type="button" class="forgot-password">¿Has olvidado la contraseña?</button>
-			</div>
-		</form>
-	</div>
-</div>
-</body>
-</html>
